[utils/database/logger] Gate non-wasm code behind //go:build !js
