+++ logger/types_js.go
@@ -1,0 +1,5 @@
+//go:build js
+
+package logger
+
+type Handler interface{}
