+++ utils/unarchive.go
@@ -1,3 +1,5 @@
+//go:build !js
+
 package utils
 
 import (
